Make `HouseScene::current_rotation` an int
To match `Tile`, make the `current_rotation` in the `HouseScene`
an int type

+++ headers/HouseScene.h
@@ -13,7 +13,7 @@ class HouseScene {
     // The current mouse house grid pos
     sf::Vector2i current_mouse_grid_position;
 
-    float current_rotation;
+    int current_rotation;
 
     bool editor_enabled;
